requirements.txt
absl-py==0.12.0
Adafruit-Blinka==6.10.0
adafruit-circuitpython-busdevice==5.0.6
adafruit-circuitpython-motor==3.3.0
adafruit-circuitpython-pca9685==3.3.6
adafruit-circuitpython-register==1.9.5
adafruit-circuitpython-servokit==1.3.4
Adafruit-GPIO==1.0.4
Adafruit-PlatformDetect==3.13.3
Adafruit-PureIO==1.1.8
Adafruit-SSD1306==1.6.2
apt-clone==0.2.1
apturl==0.5.2
argon2-cffi==20.1.0
asn1crypto==0.24.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.6.0
bleach==3.3.0
blinker==1.4
boto3==1.17.80
botocore==1.20.80
Brlapi==0.6.6
certifi==2018.1.18
cffi==1.14.5
chardet==3.0.4
click==8.0.1
colorama==0.3.7
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.23
dataclasses==0.8
decorator==4.1.2
defer==1.0.6
defusedxml==0.7.1
distro-info===0.18ubuntu0.18.04.1
efficientnet-pytorch==0.6.3
entrypoints==0.3
feedparser==5.2.1
Flask==2.0.1
funcsigs==1.0.2
future==0.18.2
futures==3.1.1
gast==0.3.3
google-pasta==0.2.0
graphsurgeon==0.4.5
grpcio==1.38.0
h5py==2.10.0
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
importlib-metadata==4.0.1
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
jetcam==0.0.0
jetcard==0.0.0
jetracer==0.0.0
jetson-stats==3.1.0
Jetson.GPIO==2.0.16
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
-e git+https://github.com/jaybdub/jupyter_clickable_image_widget@c637f70ae7ac2b5ef6104c69894bc2660e40035c#egg=jupyter_clickable_image_widget
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab==2.2.6
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
jupyterlab-widgets==1.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.1
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==2.1.1
mistune==0.8.4
mock==3.0.5
munch==2.5.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.18.5
oauth==1.0.1
oauthlib==2.0.6
object-detection==0.1
olefile==0.45.1
onboard==1.4.1
opt-einsum==3.3.0
packaging==20.9
paho-mqtt==1.6.1
PAM==0.4.2
pandas==0.22.0
pandocfilters==1.4.3
parso==0.8.2
pbr==5.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==5.1.0
pkg_resources==0.0.0
pretrainedmodels==0.7.4
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.17.1
ptyprocess==0.7.0
pybind11==2.6.2
pycairo==1.16.2
pycocotools==2.0.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pyftdi==0.53.1
Pygments==2.9.0
PyGObject==3.26.1
PyICU==1.9.8
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.2.0
pyRFC3339==1.0
pyrsistent==0.17.3
pyserial==3.5
python-apt==1.6.6
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2018.3
pyusb==1.1.1
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
pyzmq==22.1.0
qtconsole==5.1.0
QtPy==1.9.0
requests==2.18.4
requests-unixsocket==0.1.5
s3transfer==0.4.2
scikit-learn==0.24.2
scipy==0.19.1
SecretStorage==2.3.1
segmentation-models-pytorch==0.1.3
Send2Trash==1.5.0
simplejson==3.13.2
six==1.11.0
slim==0.1
spidev==3.5
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tensorboard==1.15.0
tensorflow==1.15.5+nv21.5
tensorflow-estimator==1.15.1
tensorrt==7.1.3.0
termcolor==1.1.0
terminado==0.10.0
testpath==0.5.0
testresources==2.0.1
threadpoolctl==2.1.0
timm==0.3.2
torch @ file:///home/jetson/torch-1.6.0-cp36-cp36m-linux_aarch64.whl
torch2trt==0.2.0
torchvision==0.7.0a0+78ed10c
tornado==6.1
tqdm==4.61.0
traitlets @ git+https://github.com/ipython/traitlets@dead2b8cdde5913572254cf6dc70b5a6065b86f8
typing-extensions==3.10.0.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
uff==0.6.9
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.26.4
urwid==2.0.1
wadllib==1.3.2
wcwidth==0.2.5
webencodings==0.5
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xkit==0.0.0
youtube_dl==2018.3.14
zipp==3.4.1
zope.interface==4.3.2
